feature gate the evm stuff
