+++ test-no-std/Cargo.toml
@@ -13,3 +13,6 @@ sgx-runtime = { path = "../runtime", default-features = false }
 sp-io = { path = "../substrate-sgx/sp-io", default-features = false, features = ["disable_oom", "disable_panic_handler", "disable_allocator", "sgx"] }
 sp-application-crypto = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.24", default-features = false, features = ["full_crypto"] }
 sp-core = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.24", default-features = false, features = ["full_crypto"] }
+
+[features]
+evm = ["sgx-runtime/evm"]
